requirements.txt
Flask
Flask_SQLAlchemy
Flask_Login
Flask_Bcrypt
requests
google_generativeai
python_dotenv
Flask_Caching
flask_mail